requirements.txt
streamlit
python-pptx
openai
pandas
python-dotenv

